# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"tokenVersion": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"messagesSent": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"messagesReceived": {
						"items": {
							"$ref": "#/components/schemas/Message"
						},
						"type": "array"
					},
					"profilePictureId": {
						"type": "number",
						"format": "double"
					},
					"profilePicture": {
						"$ref": "#/components/schemas/Picture"
					}
				},
				"required": [
					"id",
					"email",
					"password",
					"roleId",
					"role",
					"tokenVersion",
					"messagesSent",
					"messagesReceived"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"role",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"senderId": {
						"type": "number",
						"format": "double"
					},
					"receiverId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"delivered": {
						"type": "boolean"
					},
					"read": {
						"type": "boolean"
					},
					"sender": {
						"$ref": "#/components/schemas/User"
					},
					"receiver": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"content",
					"senderId",
					"receiverId",
					"createdAt",
					"delivered",
					"read",
					"sender",
					"receiver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Listing": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"statusId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/ListingStatus"
					},
					"personId": {
						"type": "number",
						"format": "double"
					},
					"person": {
						"$ref": "#/components/schemas/Person"
					},
					"pictures": {
						"items": {
							"$ref": "#/components/schemas/Picture"
						},
						"type": "array"
					},
					"faultCategoryId": {
						"type": "number",
						"format": "double"
					},
					"faultCategory": {
						"$ref": "#/components/schemas/FaultCategory"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					},
					"cityId": {
						"type": "number",
						"format": "double"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"createdAt",
					"updatedAt",
					"statusId",
					"status",
					"personId",
					"person",
					"pictures",
					"faultCategoryId",
					"faultCategory",
					"offers",
					"cityId",
					"city"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListingStatus": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"status",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Person": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/Review"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"userId",
					"user",
					"createdAt",
					"updatedAt",
					"reviews",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Service": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"oib": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"cityId": {
						"type": "number",
						"format": "double"
					},
					"city": {
						"$ref": "#/components/schemas/City"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/Review"
						},
						"type": "array"
					},
					"faultCategories": {
						"items": {
							"$ref": "#/components/schemas/FaultCategory"
						},
						"type": "array"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"oib",
					"phone",
					"address",
					"cityId",
					"city",
					"userId",
					"user",
					"createdAt",
					"updatedAt",
					"reviews",
					"offers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"City": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"services": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"postalCode",
					"services",
					"listings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Review": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"authorId": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"$ref": "#/components/schemas/Person"
					},
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					}
				},
				"required": [
					"id",
					"content",
					"score",
					"createdAt",
					"updatedAt",
					"authorId",
					"author",
					"serviceId",
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FaultCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"parent": {
						"$ref": "#/components/schemas/FaultCategory"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/FaultCategory"
						},
						"type": "array"
					},
					"listings": {
						"items": {
							"$ref": "#/components/schemas/Listing"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Offer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"service": {
						"$ref": "#/components/schemas/Service"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"statusId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/OfferStatus"
					},
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"listing": {
						"$ref": "#/components/schemas/Listing"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"price",
					"serviceId",
					"service",
					"createdAt",
					"updatedAt",
					"statusId",
					"status",
					"listingId",
					"listing"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OfferStatus": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"offers": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"status",
					"offers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Picture": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"cloudinaryId": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"listingId": {
						"type": "number",
						"format": "double"
					},
					"listing": {
						"$ref": "#/components/schemas/Listing"
					}
				},
				"required": [
					"id",
					"name",
					"cloudinaryId",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPersonPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IOfferPayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "number",
						"format": "double"
					},
					"listingId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"description",
					"price",
					"serviceId",
					"listingId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOfferPaginatedResult": {
				"properties": {
					"current_page": {
						"type": "number",
						"format": "double"
					},
					"total_pages": {
						"type": "number",
						"format": "double"
					},
					"per_page": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Offer"
						},
						"type": "array"
					}
				},
				"required": [
					"current_page",
					"total_pages",
					"per_page",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/people": {
			"get": {
				"operationId": "GetPeople",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Person"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePerson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Person"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPersonPayload"
							}
						}
					}
				}
			}
		},
		"/people/{id}": {
			"get": {
				"operationId": "GetPerson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Person"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePerson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Person"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPersonPayload"
							}
						}
					}
				}
			}
		},
		"/people/user/{id}": {
			"get": {
				"operationId": "GetPersonByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/people/data/{id}": {
			"get": {
				"operationId": "GetDashboardData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/offers": {
			"get": {
				"operationId": "GetOffers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Offer"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Offer"
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOfferPayload"
							}
						}
					}
				}
			}
		},
		"/offers/{id}": {
			"put": {
				"operationId": "UpdateOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Offer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOfferPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Offer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/offers/id/{id}": {
			"get": {
				"operationId": "GetOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Offer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/offers/active/{id}": {
			"get": {
				"operationId": "GetActiveOffers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOfferPaginatedResult"
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/offers/history/{id}": {
			"get": {
				"operationId": "GetHistoryOffers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOfferPaginatedResult"
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/offers/accept/{id}": {
			"post": {
				"operationId": "AcceptOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Offer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/offers/decline/{id}": {
			"post": {
				"operationId": "DeclineOffer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Offer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Offer"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}